edited common

+++ src/main/scala/objektwerks/Common.scala
@@ -9,7 +9,7 @@ final case class DoubleRange(low: Double, high: Double) derives JsonSupport:
   def avg: Double = format( (low + high) / 2 )
 
 final case class IntRange(low: Int, high: Int) derives JsonSupport:
-  def avg: Int = ( (low + high) / 2 ).toInt
+  def avg: Int = (low + high) / 2
 
 final case class TempDuration(tempRange: IntRange, duration: Int, unit: UoT) derives JsonSupport
